Add files via upload
The blood donor application was created with the intention to help folks in the local areas have access to donors and recipients without much stress. Each page includes data collection of who is ready to donate blood, the register, and inputting their contact so they can be easily reached whenever needed. Not just only for donors, but also if you wish to receive blood there is also a page to fill in your details with blood type age and all the necessary information. A work in progress.


The GVA customer form, as the name implies, is a functioning customer form for an organization to help keep the records of customers that book spaces. with the help of the canvas model app and database(dataverse), the information and data of each client are secured.